+++ dataforge-io/src/commonTest/kotlin/space/kscience/dataforge/io/EnvelopeFormatTest.kt
@@ -1,5 +1,6 @@
 package space.kscience.dataforge.io
 
+import io.ktor.utils.io.core.ByteReadPacket
 import io.ktor.utils.io.core.readDouble
 import io.ktor.utils.io.core.writeDouble
 import kotlin.test.Test
@@ -9,10 +10,10 @@ import kotlin.test.assertEquals
 class EnvelopeFormatTest {
     val envelope = Envelope {
         type = "test.format"
-        meta{
+        meta {
             "d" put 22.2
         }
-        data{
+        data {
             writeDouble(22.2)
 //            repeat(2000){
 //                writeInt(it)
@@ -21,12 +22,12 @@ class EnvelopeFormatTest {
     }
 
     @Test
-    fun testTaggedFormat(){
+    fun testTaggedFormat() {
         TaggedEnvelopeFormat.run {
             val byteArray = writeToByteArray(envelope)
             //println(byteArray.decodeToString())
             val res = readFromByteArray(byteArray)
-            assertEquals(envelope.meta,res.meta)
+            assertEquals(envelope.meta, res.meta)
             val double = res.data?.read {
                 readDouble()
             }
@@ -35,12 +36,14 @@ class EnvelopeFormatTest {
     }
 
     @Test
-    fun testTaglessFormat(){
+    fun testTaglessFormat() {
         TaglessEnvelopeFormat.run {
             val byteArray = writeToByteArray(envelope)
             //println(byteArray.decodeToString())
+            val partial = readPartial(ByteReadPacket(byteArray))
+            assertEquals(8, partial.dataSize?.toInt())
             val res = readFromByteArray(byteArray)
-            assertEquals(envelope.meta,res.meta)
+            assertEquals(envelope.meta, res.meta)
             val double = res.data?.read {
                 readDouble()
             }